+++ NEWS.md
@@ -1,3 +1,15 @@
+# ompr.roi 0.8.0
+
+## Breaking changes
+
+* The package exports column and row duals of LPs. Extracting duals might be solver dependent. Only tested with GLPK. This requires version `0.8` of ompr.
+* Minimum supported version is `3.4.0` since `slam` also requires `>= 3.4`
+
+## New features
+
+* `ompr.roi` works with other `ompr` backends as well.
+* `ompr.roi` always returns a solution, even if the solution status is not optimal. It extracts the `ROI` solution with `force=TRUE`.
+
 # ompr.roi 0.7.0
 
 ## Major breaking changes